@@ -6,28 +6,34 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.21.0",
-    "@tanstack/react-query": "^4.29.5",
-    "@tanstack/react-query-devtools": "^4.29.6",
-    "date-fns": "^2.30.0",
+    "@supabase/supabase-js": "^2.45.2",
+    "@tanstack/react-query": "^4.36.1",
+    "@tanstack/react-query-devtools": "^4.36.1",
+    "date-fns": "^3.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.43.9",
+    "react-error-boundary": "^4.0.13",
+    "react-hook-form": "^7.53.0",
     "react-hot-toast": "^2.4.1",
-    "react-icons": "^4.8.0",
-    "react-router-dom": "^6.11.1",
-    "styled-components": "^5.3.10"
+    "react-icons": "^5.3.0",
+    "react-router-dom": "^6.26.1",
+    "recharts": "^2.12.7",
+    "styled-components": "^6.1.12"
   },
   "devDependencies": {
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "@vitejs/plugin-react": "^3.1.0",
-    "eslint": "^8.39.0",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.57.0",
     "eslint-config-react-app": "^7.0.1",
-    "vite": "^4.2.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "vite": "^4.4.5",
     "vite-plugin-eslint": "^1.8.1"
   }
 }
