@@ -1,18 +1,34 @@
 import { useState } from "react";
 import Button from "../../ui/Button";
 import Form from "../../ui/Form";
-import FormRow from "../../ui/FormRow";
 import Input from "../../ui/Input";
+import FormRowVertical from "../../ui/FormRowVertical";
+import { useLogin } from "./useLogin";
+import SpinnerMini from "../../ui/SpinnerMini";
 
 function LoginForm() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+  const [email, setEmail] = useState("praveen1220raj@gmail.com");
+  const [password, setPassword] = useState("12345678");
+  const { login, isLoading } = useLogin();
 
-  function handleSubmit() {}
+  function handleSubmit(e) {
+    e.preventDefault();
+
+    if (!email || !password) return;
+    login(
+      { email, password },
+      {
+        onSettled: () => {
+          setEmail("");
+          setPassword("");
+        },
+      }
+    );
+  }
 
   return (
     <Form onSubmit={handleSubmit}>
-      <FormRow label="Email address" orientation="vertical">
+      <FormRowVertical label="Email address">
         <Input
           type="email"
           id="email"
@@ -20,20 +36,24 @@ function LoginForm() {
           autoComplete="username"
           value={email}
           onChange={(e) => setEmail(e.target.value)}
+          disabled={isLoading}
         />
-      </FormRow>
-      <FormRow label="Password" orientation="vertical">
+      </FormRowVertical>
+      <FormRowVertical label="Password">
         <Input
           type="password"
           id="password"
           autoComplete="current-password"
           value={password}
           onChange={(e) => setPassword(e.target.value)}
+          disabled={isLoading}
         />
-      </FormRow>
-      <FormRow orientation="vertical">
-        <Button size="large">Login</Button>
-      </FormRow>
+      </FormRowVertical>
+      <FormRowVertical>
+        <Button size="large" disabled={isLoading}>
+          {!isLoading ? "log in" : <SpinnerMini />}
+        </Button>
+      </FormRowVertical>
     </Form>
   );
 }
