@@ -1,65 +1,59 @@
-import { useForm } from 'react-hook-form';
-import Button from 'ui/Button';
-import Form from 'ui/Form';
-import FormRow from 'ui/FormRow';
-import Input from 'ui/Input';
-import { useUpdateUser } from './useUpdateUser';
+import { useForm } from "react-hook-form";
+import Button from "../../ui/Button";
+import Form from "../../ui/Form";
+import FormRow from "../../ui/FormRow";
+import Input from "../../ui/Input";
+
+import { useUpdateUser } from "./useUpdateUser";
 
 function UpdatePasswordForm() {
   const { register, handleSubmit, formState, getValues, reset } = useForm();
   const { errors } = formState;
 
-  const { mutate: updateUser, isLoading: isUpdating } = useUpdateUser();
+  const { updateUser, isUpdating } = useUpdateUser();
 
   function onSubmit({ password }) {
-    updateUser({ password }, { onSuccess: () => reset() });
+    updateUser({ password }, { onSuccess: reset });
   }
-
-  function handleReset(e) {
-    // e.preventDefault();
-    reset();
-  }
-
   return (
     <Form onSubmit={handleSubmit(onSubmit)}>
       <FormRow
-        label='Password (min 8 characters)'
+        label="Password (min 8 characters)"
         error={errors?.password?.message}
       >
         <Input
-          type='password'
-          id='password'
-          // this makes the form better for password managers
-          autoComplete='current-password'
+          type="password"
+          id="password"
+          autoComplete="current-password"
           disabled={isUpdating}
-          {...register('password', {
-            required: 'This field is required',
+          {...register("password", {
+            required: "This field is required",
             minLength: {
               value: 8,
-              message: 'Password needs a minimum of 8 characters',
+              message: "Password needs a minimum of 8 characters",
             },
           })}
         />
       </FormRow>
 
       <FormRow
-        label='Confirm password'
+        label="Confirm password"
         error={errors?.passwordConfirm?.message}
       >
         <Input
-          type='password'
-          autoComplete='new-password'
-          id='passwordConfirm'
+          type="password"
+          autoComplete="new-password"
+          id="passwordConfirm"
           disabled={isUpdating}
-          {...register('passwordConfirm', {
-            required: 'This field is required',
+          {...register("passwordConfirm", {
+            required: "This field is required",
             validate: (value) =>
-              getValues().password === value || 'Passwords need to match',
+              getValues().password === value || "Passwords need to match",
           })}
         />
       </FormRow>
       <FormRow>
-        <Button onClick={handleReset} type='reset' variation='secondary'>
+        <Button onClick={reset} type="reset" variation="secondary">
           Cancel
         </Button>
         <Button disabled={isUpdating}>Update password</Button>
