@@ -1,73 +1,75 @@
-import { useUser } from 'features/authentication/useUser';
-import { useState } from 'react';
-import Button from 'ui/Button';
-import FileInput from 'ui/FileInput';
-import Form from 'ui/Form';
-import FormRow from 'ui/FormRow';
-import Input from 'ui/Input';
-import { useUpdateUser } from './useUpdateUser';
+import { useState } from "react";
+
+import Button from "../../ui/Button";
+import FileInput from "../../ui/FileInput";
+import Form from "../../ui/Form";
+import FormRow from "../../ui/FormRow";
+import Input from "../../ui/Input";
+
+import { useUser } from "./useUser";
+import { useUpdateUser } from "./useUpdateUser";
 
 function UpdateUserDataForm() {
-  // We don't need the loading state
+  // We don't need the loading state, and can immediately use the user data, because we know that it has already been loaded at this point
   const {
     user: {
       email,
       user_metadata: { fullName: currentFullName },
     },
   } = useUser();
+  const { updateUser, isUpdating } = useUpdateUser();
 
   const [fullName, setFullName] = useState(currentFullName);
   const [avatar, setAvatar] = useState(null);
 
-  const { mutate: updateUser, isLoading: isUpdating } = useUpdateUser();
-
   function handleSubmit(e) {
     e.preventDefault();
     if (!fullName) return;
-
     updateUser(
       { fullName, avatar },
       {
-        onSuccess: () => {
+        onSettled: () => {
           setAvatar(null);
-          // Resetting form using .reset() that's available on all HTML form elements, otherwise the old filename will stay displayed in the UI
           e.target.reset();
         },
       }
     );
   }
 
-  function handleCancel(e) {
-    // We don't even need preventDefault because this button was designed to reset the form (remember, it has the HTML attribute 'reset')
-    setFullName(currentFullName);
+  function handleCancel() {
     setAvatar(null);
+    setFullName(currentFullName);
   }
 
   return (
     <Form onSubmit={handleSubmit}>
-      <FormRow label='Email address'>
+      <FormRow label="Email address">
         <Input value={email} disabled />
       </FormRow>
-      <FormRow label='Full name'>
+      <FormRow label="Full name">
         <Input
-          type='text'
+          disabled={isUpdating}
+          type="text"
           value={fullName}
           onChange={(e) => setFullName(e.target.value)}
-          disabled={isUpdating}
-          id='fullName'
+          id="fullName"
         />
       </FormRow>
-      <FormRow label='Avatar image'>
+      <FormRow label="Avatar image">
         <FileInput
           disabled={isUpdating}
-          id='avatar'
-          accept='image/*'
+          id="avatar"
+          accept="image/*"
           onChange={(e) => setAvatar(e.target.files[0])}
-          // We should also validate that it's actually an image, but never mind
         />
       </FormRow>
       <FormRow>
-        <Button onClick={handleCancel} type='reset' variation='secondary'>
+        <Button
+          onClick={handleCancel}
+          disabled={isUpdating}
+          type="reset"
+          variation="secondary"
+        >
           Cancel
         </Button>
         <Button disabled={isUpdating}>Update account</Button>
