@@ -1,46 +1,21 @@
-// import styled from 'styled-components';
-import BookingRow from 'features/bookings/BookingRow';
-import Spinner from 'ui/Spinner';
-import Table from 'ui/Table';
-import { useBookings } from 'features/bookings/useBookings';
-import Menus from 'ui/Menus';
-import Pagination from 'ui/Pagination';
-import Empty from 'ui/Empty';
-
-// v2
-// Right now this is not really reusable... But we will want to use a similar table for guests as well, but with different columns. ALSO, right now we are defining these columns in BOTH the TableHeader and the BookingRow, which is not good at all. Instead, it would be much better to simply pass the columns into the Table, and the table would give access to the columns to both the header and row. So how can we do that? Well we can again use a compound component! We don't HAVE to do it like this, there's a million ways to implement a table, also without CSS Grid, but this is what I chose
-
-// v1
-// const TableHeader = styled.header`
-//   display: grid;
-//   grid-template-columns: 0.6fr 1.8fr 2.2fr 1fr 1fr 1fr;
-//   column-gap: 2.4rem;
-//   align-items: center;
-
-//   background-color: var(--color-grey-50);
-//   border-bottom: 1px solid var(--color-grey-100);
-//   text-transform: uppercase;
-//   letter-spacing: 0.4px;
-//   font-weight: 600;
-//   color: var(--color-grey-600);
-//   padding: 1.6rem 2.4rem;
-// `;
-
-// We want each table row to have a menu, and we only want one of them to be open at the same time. We also want this functionality to be reusable. We could add a openID state here to the table, but that wouldn't really be reusable... The best way is to use a compound component
+import BookingRow from "./BookingRow";
+import Table from "../../ui/Table";
+import Menus from "../../ui/Menus";
+import Empty from "../../ui/Empty";
+import { useBookings } from "./useBookings";
+import Spinner from "../../ui/Spinner";
+import Pagination from "../../ui/Pagination";
 
 function BookingTable() {
-  const { bookings, count, isLoading } = useBookings();
+  const { bookings, isLoading, count } = useBookings();
 
   if (isLoading) return <Spinner />;
-  if (!bookings) return <Empty resource={'bookings'} />;
 
-  // VIDEO stupid JS bug, just an example of course
-  // null.toUpperCase();
+  if (!bookings.length) return <Empty resourceName="bookings" />;
 
   return (
     <Menus>
-      {/* A beautiful API we created here! We could even have defined the widths on the columns in the table header individually, but this keeps it simpler, and I also really like it */}
-      <Table columns='0.6fr 2fr 2.4fr 1.4fr 1fr 3.2rem'>
+      <Table columns="0.6fr 2fr 2.4fr 1.4fr 1fr 3.2rem">
         <Table.Header>
           <div>Cabin</div>
           <div>Guest</div>
@@ -50,18 +25,12 @@ function BookingTable() {
           <div></div>
         </Table.Header>
 
-        {/* {bookings.map((booking) => (
-            <BookingRow key={booking.id} booking={booking} />
-          ))} */}
-
-        {/* Render props! */}
         <Table.Body
           data={bookings}
           render={(booking) => (
             <BookingRow key={booking.id} booking={booking} />
           )}
         />
-
         <Table.Footer>
           <Pagination count={count} />
         </Table.Footer>
@@ -69,8 +38,5 @@ function BookingTable() {
     </Menus>
   );
 }
-
-// We could create yet another layer of abstraction on top of this. We could call this component just <Results>, like: Results({data, count, isLoading, columns, rowComponent}). Then <BookingTable> and ALL other tables would simply call that.
-// BUT, creating more abstractions also has a cost! More things to remember, more complex codebase to understand. Sometimes it's okay to just copy and paste instead of creating abstractions
 
 export default BookingTable;
