@@ -1,30 +1,29 @@
-import SortBy from 'ui/SortBy';
-import Filter from 'ui/Filter';
-import TableOperations from 'ui/TableOperations';
+import SortBy from "../../ui/SortBy";
+import Filter from "../../ui/Filter";
+import TableOperations from "../../ui/TableOperations";
 
 function BookingTableOperations() {
   return (
     <TableOperations>
-      {/* We could do these two as compound components as well, but let's keep it simple, and let's also explore different ways of achieving the same thing */}
       <Filter
-        filterField='status'
+        filterField="status"
         options={[
-          { value: 'all', label: 'All' },
-          { value: 'checked-out', label: 'Checked out' },
-          { value: 'checked-in', label: 'Checked in' },
-          { value: 'unconfirmed', label: 'Unconfirmed' },
+          { value: "all", label: "All" },
+          { value: "checked-out", label: "Checked out" },
+          { value: "checked-in", label: "Checked in" },
+          { value: "unconfirmed", label: "Unconfirmed" },
         ]}
       />
 
       <SortBy
         options={[
-          { value: 'startDate-desc', label: 'Sort by date (recent first)' },
-          { value: 'startDate-asc', label: 'Sort by date (earlier first)' },
+          { value: "startDate-desc", label: "Sort by date (recent first)" },
+          { value: "startDate-asc", label: "Sort by date (earlier first)" },
           {
-            value: 'totalPrice-desc',
-            label: 'Sort by amount (high first)',
+            value: "totalPrice-desc",
+            label: "Sort by amount (high first)",
           },
-          { value: 'totalPrice-asc', label: 'Sort by amount (low first)' },
+          { value: "totalPrice-asc", label: "Sort by amount (low first)" },
         ]}
       />
     </TableOperations>
