@@ -1,5 +1,4 @@
 import styled from "styled-components";
-
 import Spinner from "../../ui/Spinner";
 import CabinRow from "./CabinRow";
 import { useCabins } from "./useCabins";
@@ -30,11 +29,10 @@ const TableHeader = styled.header`
 
 function CabinTable() {
   const { isLoading, cabins } = useCabins();
-
   if (isLoading) return <Spinner />;
 
   return (
-    <Table role="table">
+    <Table>
       <TableHeader role="row">
         <div></div>
         <div>Cabin</div>
