@@ -1,5 +1,3 @@
-import { useForm } from "react-hook-form";
-
 import Input from "../../ui/Input";
 import Form from "../../ui/Form";
 import Button from "../../ui/Button";
@@ -7,30 +5,31 @@ import FileInput from "../../ui/FileInput";
 import Textarea from "../../ui/Textarea";
 import FormRow from "../../ui/FormRow";
 
+import { useForm } from "react-hook-form";
 import { useCreateCabin } from "./useCreateCabin";
 import { useEditCabin } from "./useEditCabin";
 
 function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
-  const { isCreating, createCabin } = useCreateCabin();
-  const { isEditing, editCabin } = useEditCabin();
-  const isWorking = isCreating || isEditing;
-
   const { id: editId, ...editValues } = cabinToEdit;
   const isEditSession = Boolean(editId);
 
   const { register, handleSubmit, reset, getValues, formState } = useForm({
     defaultValues: isEditSession ? editValues : {},
   });
+
   const { errors } = formState;
 
+  const { isCreating, createCabin } = useCreateCabin();
+  const { isEditing, editCabin } = useEditCabin();
+
+  const isWorking = isCreating || isEditing;
   function onSubmit(data) {
     const image = typeof data.image === "string" ? data.image : data.image[0];
-
     if (isEditSession)
       editCabin(
         { newCabinData: { ...data, image }, id: editId },
         {
-          onSuccess: (data) => {
+          onSuccess: () => {
             reset();
             onCloseModal?.();
           },
@@ -40,7 +39,7 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
       createCabin(
         { ...data, image: image },
         {
-          onSuccess: (data) => {
+          onSuccess: () => {
             reset();
             onCloseModal?.();
           },
@@ -48,23 +47,21 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
       );
   }
 
-  function onError(errors) {
-    // console.log(errors);
+  function onError(data) {
+    console.error(data);
   }
 
   return (
     <Form
-      onSubmit={handleSubmit(onSubmit, onError)}
       type={onCloseModal ? "modal" : "regular"}
+      onSubmit={handleSubmit(onSubmit, onError)}
     >
       <FormRow label="Cabin name" error={errors?.name?.message}>
         <Input
+          disabled={isWorking}
           type="text"
           id="name"
-          disabled={isWorking}
-          {...register("name", {
-            required: "This field is required",
-          })}
+          {...register("name", { required: "This field is Required" })}
         />
       </FormRow>
 
@@ -74,7 +71,7 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
           id="maxCapacity"
           disabled={isWorking}
           {...register("maxCapacity", {
-            required: "This field is required",
+            required: "This field is Required",
             min: {
               value: 1,
               message: "Capacity should be at least 1",
@@ -89,10 +86,10 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
           id="regularPrice"
           disabled={isWorking}
           {...register("regularPrice", {
-            required: "This field is required",
+            required: "This field is Required",
             min: {
               value: 1,
-              message: "Capacity should be at least 1",
+              message: "Price should be at least 1",
             },
           })}
         />
@@ -102,13 +99,17 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
         <Input
           type="number"
           id="discount"
-          disabled={isWorking}
           defaultValue={0}
+          disabled={isWorking}
           {...register("discount", {
-            required: "This field is required",
+            required: "This field is Required",
             validate: (value) =>
               value <= getValues().regularPrice ||
-              "Discount should be less than regular price",
+              "Discount Should be less than regular price",
+            min: {
+              value: 0,
+              message: "Discount should be greater than / equal to 0",
+            },
           })}
         />
       </FormRow>
@@ -118,22 +119,21 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
         error={errors?.description?.message}
       >
         <Textarea
+          disabled={isWorking}
           type="number"
           id="description"
           defaultValue=""
-          disabled={isWorking}
-          {...register("description", {
-            required: "This field is required",
-          })}
+          {...register("description", { required: "This field is Required" })}
         />
       </FormRow>
 
       <FormRow label="Cabin photo">
         <FileInput
+          disabled={isWorking}
           id="image"
           accept="image/*"
           {...register("image", {
-            required: isEditSession ? false : "This field is required",
+            required: isEditSession ? false : "This field is Required",
           })}
         />
       </FormRow>
@@ -141,14 +141,15 @@ function CreateCabinForm({ cabinToEdit = {}, onCloseModal }) {
       <FormRow>
         {/* type is an HTML attribute! */}
         <Button
+          onClick={() => onCloseModal?.()}
+          disabled={isWorking}
           variation="secondary"
           type="reset"
-          onClick={() => onCloseModal?.()}
         >
           Cancel
         </Button>
         <Button disabled={isWorking}>
-          {isEditSession ? "Edit cabin" : "Create new cabin"}
+          {isEditSession ? "Edit cabin" : "Create new Cabin"}
         </Button>
       </FormRow>
     </Form>
