@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { toast } from "react-hot-toast";
+import toast from "react-hot-toast";
 import { deleteCabin as deleteCabinApi } from "../../services/apiCabins";
 
 export function useDeleteCabin() {
@@ -9,13 +9,11 @@ export function useDeleteCabin() {
     mutationFn: deleteCabinApi,
     onSuccess: () => {
       toast.success("Cabin successfully deleted");
-
       queryClient.invalidateQueries({
         queryKey: ["cabins"],
       });
     },
     onError: (err) => toast.error(err.message),
   });
-
   return { isDeleting, deleteCabin };
 }
