@@ -1,15 +1,17 @@
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import { box } from 'styles/styles';
-import Button from 'ui/Button';
-import Heading from 'ui/Heading';
-import Row from 'ui/Row';
-import Spinner from 'ui/Spinner';
-import TodayItem from '../dashboard/TodayItem';
-import { useActivityTodayStays } from './useActivityTodayStays';
+import styled from "styled-components";
+
+import Heading from "../../ui/Heading";
+import Row from "../../ui/Row";
+import { useTodayActivity } from "./useTodayActivity";
+import Spinner from "../../ui/Spinner";
+import TodayItem from "./TodayItem";
 
 const StyledToday = styled.div`
-  ${box}
+  /* Box */
+  background-color: var(--color-grey-0);
+  border: 1px solid var(--color-grey-100);
+  border-radius: var(--border-radius-md);
+
   padding: 3.2rem;
   display: flex;
   flex-direction: column;
@@ -37,22 +39,18 @@ const NoActivity = styled.p`
   margin-top: 0.8rem;
 `;
 
-function Today() {
-  const { isLoading, stays } = useActivityTodayStays();
-
+function TodayActivity() {
+  const { isLoading, activities } = useTodayActivity();
   return (
     <StyledToday>
-      <Row type='horizontal'>
-        <Heading type='h2'>Today</Heading>
-        {/* Through the 'as' props, we make the button Polymorphic! Built-in into styled components. The polymorphic component pattern comes in handy when we need flexibility on the rendered HTML element. */}
-        {/* id of -1 means there is no ID, which means a new booking will be made for a new guest */}
+      <Row type="horizontal">
+        <Heading as="h2">Today</Heading>
       </Row>
-
       {!isLoading ? (
-        stays?.length > 0 ? (
+        activities?.length > 0 ? (
           <TodayList>
-            {stays.map((stay) => (
-              <TodayItem key={stay.id} stay={stay} />
+            {activities.map((activity) => (
+              <TodayItem key={activity.id} activity={activity} />
             ))}
           </TodayList>
         ) : (
@@ -65,43 +63,4 @@ function Today() {
   );
 }
 
-export default Today;
-
-const OLDdata = [
-  {
-    id: 1,
-    status: 'unconfirmed',
-    guests: { fullName: 'Jonas Schmedtmann' },
-    numNights: 6,
-  },
-  {
-    id: 2,
-    status: 'unconfirmed',
-    guests: { fullName: 'Steven Miller' },
-    numNights: 1,
-  },
-  {
-    id: 3,
-    status: 'checked-in',
-    guests: { fullName: 'John Smith' },
-    numNights: 3,
-  },
-  {
-    id: 4,
-    status: 'unconfirmed',
-    guests: { fullName: 'Marta Schmedtmann' },
-    numNights: 14,
-  },
-  {
-    id: 5,
-    status: 'checked-in',
-    guests: { fullName: 'Miguel Silva' },
-    numNights: 5,
-  },
-  {
-    id: 6,
-    status: 'checked-in',
-    guests: { fullName: 'Mary Williams' },
-    numNights: 4,
-  },
-];
+export default TodayActivity;
