@@ -1,8 +1,11 @@
-import styled from 'styled-components';
-import { box } from 'styles/styles';
+import styled from "styled-components";
 
 const DashboardBox = styled.div`
-  ${box}
+  /* Box */
+  background-color: var(--color-grey-0);
+  border: 1px solid var(--color-grey-100);
+  border-radius: var(--border-radius-md);
+
   padding: 3.2rem;
 
   display: flex;
