@@ -1,13 +1,13 @@
-import styled from 'styled-components';
+import styled from "styled-components";
+import Spinner from "../../ui/Spinner";
 
-import DurationChart from 'features/dashboard/DurationChart';
-import SalesChart from 'features/dashboard/SalesChart';
-import Stats from 'features/dashboard/Stats';
-import TodayActivity from 'features/check-in-out/TodayActivity';
-import { useRecentBookings } from 'features/dashboard/useRecentBookings';
-import Spinner from 'ui/Spinner';
-import { useRecentStays } from './useRecentStays';
-import { useCabins } from 'features/cabins/useCabins';
+import { useCabins } from "../cabins/useCabins";
+import { useRecentBookings } from "./useRecentBookings";
+import { useRecentStays } from "./useRecentStays";
+import Stats from "./Stats";
+import SalesChart from "./SalesChart";
+import DurationChart from "./DurationChart";
+import TodayActivity from "../check-in-out/TodayActivity";
 
 const StyledDashboardLayout = styled.div`
   display: grid;
@@ -16,17 +16,10 @@ const StyledDashboardLayout = styled.div`
   gap: 2.4rem;
 `;
 
-/*
-We need to distinguish between two types of data here:
-1) BOOKINGS: the actual sales. For example, in the last 30 days, the hotel might have sold 10 bookings online, but these guests might only arrive and check in in the far future (or not, as booking also happen on-site)
-2) STAYS: the actual check-in of guests arriving for their bookings. We can identify stays by their startDate, together with a status of either 'checked-in' (for current stays) or 'checked-out' (for past stays)
-*/
-
 function DashboardLayout() {
-  const { isLoading: isLoading1, bookings, numDays } = useRecentBookings();
-  const { isLoading: isLoading2, confirmedStays } = useRecentStays();
-  const { isLoading: isLoading3, cabins } = useCabins();
-
+  const { isLoading: isLoading1, bookings } = useRecentBookings();
+  const { isLoading: isLoading2, confirmedStays, numDays } = useRecentStays();
+  const { cabins, isLoading: isLoading3 } = useCabins();
   if (isLoading1 || isLoading2 || isLoading3) return <Spinner />;
 
   return (
@@ -37,8 +30,11 @@ function DashboardLayout() {
         numDays={numDays}
         cabinCount={cabins.length}
       />
+
       <TodayActivity />
+
       <DurationChart confirmedStays={confirmedStays} />
+
       <SalesChart bookings={bookings} numDays={numDays} />
     </StyledDashboardLayout>
   );
