@@ -1,53 +1,49 @@
-import {
-  HiOutlineBriefcase,
-  HiOutlineCalendarDays,
-  HiOutlineBanknotes,
-  HiOutlineChartBar,
-} from 'react-icons/hi2';
-import { formatCurrency } from 'utils/helpers';
-import Stat from './Stat';
+import { HiOutlineBriefcase, HiOutlineChartBar } from "react-icons/hi";
+import Stat from "./Stat";
+import { HiOutlineBanknotes, HiOutlineCalendarDays } from "react-icons/hi2";
+import { formatCurrency } from "../../utils/helpers";
 
 function Stats({ bookings, confirmedStays, numDays, cabinCount }) {
-  // Stat 1)
+  // 1.
   const numBookings = bookings.length;
 
-  // Stat 2)
+  // 2.
   const sales = bookings.reduce((acc, cur) => acc + cur.totalPrice, 0);
 
-  // Stat 3)
+  // 3.
   const checkins = confirmedStays.length;
 
-  // Stat 4)
-  // We will use a trick to calculate occupancy rate. It's not 100% accurate, but we want to keep it simple. We know we can have a total of 'numDays * cabinCount' days to occupy, and we also know how many days were actually booked. From this, we can compute the percentage
+  // 4.
   const occupation =
     confirmedStays.reduce((acc, cur) => acc + cur.numNights, 0) /
     (numDays * cabinCount);
+  // num checked in nights / all available nights (num days * num cabins)
 
   return (
     <>
       <Stat
         icon={<HiOutlineBriefcase />}
-        title='Bookings'
+        title="Booking"
         value={numBookings}
-        color='blue'
+        color="blue"
       />
       <Stat
         icon={<HiOutlineBanknotes />}
-        title='Sales'
+        title="Sales"
         value={formatCurrency(sales)}
-        color='green'
+        color="green"
       />
       <Stat
         icon={<HiOutlineCalendarDays />}
-        title='Check ins'
+        title="Check ins"
         value={checkins}
-        color='indigo'
+        color="indigo"
       />
       <Stat
         icon={<HiOutlineChartBar />}
-        title='Occupancy rate'
-        value={Math.round(occupation * 100) + '%'}
-        color='yellow'
+        title="Occupancy rate"
+        value={Math.round(occupation * 100) + "%"}
+        color="yellow"
       />
     </>
   );
