@@ -1,8 +1,8 @@
 import Form from "../../ui/Form";
 import FormRow from "../../ui/FormRow";
 import Input from "../../ui/Input";
-import Spinner from "../../ui/Spinner";
 import { useSettings } from "./useSettings";
+import Spinner from "../../ui/Spinner";
 import { useUpdateSetting } from "./useUpdateSetting";
 
 function UpdateSettingsForm() {
@@ -15,14 +15,14 @@ function UpdateSettingsForm() {
       breakfastPrice,
     } = {},
   } = useSettings();
+
   const { isUpdating, updateSetting } = useUpdateSetting();
-
   if (isLoading) return <Spinner />;
-
   function handleUpdate(e, field) {
     const { value } = e.target;
 
     if (!value) return;
+
     updateSetting({ [field]: value });
   }
 
@@ -32,12 +32,11 @@ function UpdateSettingsForm() {
         <Input
           type="number"
           id="min-nights"
-          defaultValue={minBookingLength}
           disabled={isUpdating}
+          defaultValue={minBookingLength}
           onBlur={(e) => handleUpdate(e, "minBookingLength")}
         />
       </FormRow>
-
       <FormRow label="Maximum nights/booking">
         <Input
           type="number"
@@ -47,7 +46,6 @@ function UpdateSettingsForm() {
           onBlur={(e) => handleUpdate(e, "maxBookingLength")}
         />
       </FormRow>
-
       <FormRow label="Maximum guests/booking">
         <Input
           type="number"
@@ -57,7 +55,6 @@ function UpdateSettingsForm() {
           onBlur={(e) => handleUpdate(e, "maxGuestsPerBooking")}
         />
       </FormRow>
-
       <FormRow label="Breakfast price">
         <Input
           type="number"
