@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { toast } from "react-hot-toast";
+import toast from "react-hot-toast";
 import { updateSetting as updateSettingApi } from "../../services/apiSettings";
 
 export function useUpdateSetting() {
@@ -7,12 +7,12 @@ export function useUpdateSetting() {
 
   const { mutate: updateSetting, isLoading: isUpdating } = useMutation({
     mutationFn: updateSettingApi,
+
     onSuccess: () => {
       toast.success("Setting successfully edited");
       queryClient.invalidateQueries({ queryKey: ["settings"] });
     },
     onError: (err) => toast.error(err.message),
   });
-
   return { isUpdating, updateSetting };
 }
