@@ -1,4 +1,8 @@
 import styled from "styled-components";
+import LoginForm from "../features/authentication/LoginForm";
+import Logo from "../ui/Logo";
+import Heading from "../ui/Heading";
+// import CabinTable from "../features/cabins/CabinTable";
 
 const LoginLayout = styled.main`
   min-height: 100vh;
@@ -11,7 +15,14 @@ const LoginLayout = styled.main`
 `;
 
 function Login() {
-  return <LoginLayout>Login</LoginLayout>;
+  return (
+    <LoginLayout>
+      <Logo />
+      <Heading as="h4">Log in to your account</Heading>
+      <LoginForm />
+      {/* <CabinTable /> */}
+    </LoginLayout>
+  );
 }
 
 export default Login;
