@@ -1,7 +1,13 @@
+import SignupForm from "../features/authentication/SignupForm";
 import Heading from "../ui/Heading";
 
 function NewUsers() {
-  return <Heading as="h1">Create a new user</Heading>;
+  return (
+    <>
+      <Heading as="h1">Create a new user</Heading>
+      <SignupForm />
+    </>
+  );
 }
 
 export default NewUsers;
