@@ -2,10 +2,18 @@ import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
-
   if (error) {
     console.error(error);
     throw new Error("Cabins could not be loaded");
+  }
+  return data;
+}
+
+export async function deleteCabin(id) {
+  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+  if (error) {
+    console.error(error);
+    throw new Error("Cabins could not be deleted");
   }
 
   return data;
@@ -18,6 +26,7 @@ export async function createEditCabin(newCabin, id) {
     "/",
     ""
   );
+
   const imagePath = hasImagePath
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
@@ -25,44 +34,37 @@ export async function createEditCabin(newCabin, id) {
   // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
-  // A) CREATE
-  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  // A) Create cabin
+  if (!id) query = query.insert({ ...newCabin, image: imagePath });
 
-  // B) EDIT
-  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  // B) Edit cabin
+  if (id)
+    query = query
+      .update({ ...newCabin, image: imagePath })
+      .eq("id", id)
+      .select();
 
   const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error("Cabin could not be created");
+    throw new Error("Cabin Could not be created");
   }
 
-  // 2. Upload image
+  //2. Upload image
   if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin IF there was an error uplaoding image
+  // 3.Delete the cabin If there was an error uploading image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
       "Cabin image could not be uploaded and the cabin was not created"
     );
-  }
-
-  return data;
-}
-
-export async function deleteCabin(id) {
-  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
-
-  if (error) {
-    console.error(error);
-    throw new Error("Cabin could not be deleted");
   }
 
   return data;
