@@ -1,8 +1,7 @@
 import { Outlet } from "react-router-dom";
-import Sidebar from "./Sidebar";
 import Header from "./Header";
+import Sidebar from "./Sidebar";
 import styled from "styled-components";
-
 const StyledAppLayout = styled.div`
   display: grid;
   grid-template-columns: 26rem 1fr;
@@ -15,7 +14,6 @@ const Main = styled.main`
   padding: 4rem 4.8rem 6.4rem;
   overflow: scroll;
 `;
-
 const Container = styled.div`
   max-width: 120rem;
   margin: 0 auto;
@@ -23,7 +21,6 @@ const Container = styled.div`
   flex-direction: column;
   gap: 3.2rem;
 `;
-
 function AppLayout() {
   return (
     <StyledAppLayout>
