@@ -52,14 +52,8 @@ const Button = styled.button`
   border: none;
   border-radius: var(--border-radius-sm);
   box-shadow: var(--shadow-sm);
-
   ${(props) => sizes[props.size]}
   ${(props) => variations[props.variation]}
 `;
-
-Button.defaultProps = {
-  variation: "primary",
-  size: "medium",
-};
-
+Button.defaultProps = { variation: "primary", size: "medium" };
 export default Button;
