@@ -24,6 +24,7 @@ function ConfirmDelete({ resourceName, onConfirm, disabled, onCloseModal }) {
   return (
     <StyledConfirmDelete>
       <Heading as="h3">Delete {resourceName}</Heading>
+
       <p>
         Are you sure you want to delete this {resourceName} permanently? This
         action cannot be undone.
@@ -31,12 +32,13 @@ function ConfirmDelete({ resourceName, onConfirm, disabled, onCloseModal }) {
 
       <div>
         <Button
+          onClick={onCloseModal}
           variation="secondary"
           disabled={disabled}
-          onClick={onCloseModal}
         >
           Cancel
         </Button>
+
         <Button variation="danger" disabled={disabled} onClick={onConfirm}>
           Delete
         </Button>
