@@ -1,11 +1,10 @@
 import styled from "styled-components";
 
 const Input = styled.input`
+  box-shadow: var(--shadow-sm);
   border: 1px solid var(--color-grey-300);
-  background-color: var(--color-grey-0);
   border-radius: var(--border-radius-sm);
   padding: 0.8rem 1.2rem;
-  box-shadow: var(--shadow-sm);
+  background-color: var(--color-grey-0);
 `;
-
 export default Input;
