@@ -87,13 +87,14 @@ function Toggle({ id }) {
   const { openId, close, open, setPosition } = useContext(MenusContext);
 
   function handleClick(e) {
-    const rect = e.target.closest("button").getBoundingClientRect();
-    setPosition({
-      x: window.innerWidth - rect.width - rect.x,
-      y: rect.y + rect.height + 8,
-    });
+    e.stopPropagation();
+    const { x, y, width, height } = e.target
+      .closest("button")
+      .getBoundingClientRect();
 
-    openId === "" || openId !== id ? open(id) : close();
+    setPosition({ x: window.innerWidth - width - x, y: y + height + 8 });
+
+    openId === "" || openId == id ? open(id) : close();
   }
 
   return (
@@ -105,12 +106,10 @@ function Toggle({ id }) {
 
 function List({ id, children }) {
   const { openId, position, close } = useContext(MenusContext);
-  const ref = useOutsideClick(close);
-
+  const ref = useOutsideClick(close, false);
   if (openId !== id) return null;
-
   return createPortal(
-    <StyledList position={position} ref={ref}>
+    <StyledList ref={ref} position={position}>
       {children}
     </StyledList>,
     document.body
@@ -129,7 +128,7 @@ function Button({ children, icon, onClick }) {
     <li>
       <StyledButton onClick={handleClick}>
         {icon}
-        <span>{children}</span>
+        <span> {children} </span>
       </StyledButton>
     </li>
   );
