@@ -2,14 +2,12 @@ import styled, { css } from "styled-components";
 
 const Row = styled.div`
   display: flex;
-
   ${(props) =>
     props.type === "horizontal" &&
     css`
       justify-content: space-between;
       align-items: center;
     `}
-
   ${(props) =>
     props.type === "vertical" &&
     css`
@@ -17,9 +15,5 @@ const Row = styled.div`
       gap: 1.6rem;
     `}
 `;
-
-Row.defaultProps = {
-  type: "vertical",
-};
-
+Row.defaultProps = { type: "vertical" };
 export default Row;
